Enum chalk_ir::Ty

pub enum Ty<TF: TypeFamily> {
    Apply(ApplicationTy<TF>),
    Dyn(Binders<Vec<QuantifiedWhereClause<TF>>>),
    Opaque(Binders<Vec<QuantifiedWhereClause<TF>>>),
    Projection(ProjectionTy<TF>),
    ForAll(Box<QuantifiedTy<TF>>),
    BoundVar(usize),
    InferenceVar(InferenceVar),
}

Variants
Apply(ApplicationTy<TF>)

An "application" type is one that applies the set of type arguments to some base type. For example, Vec<u32> would be "applying" the parameters [u32] to the code type Vec. This type is also used for base types like u32 (which just apply an empty list).

Dyn(Binders<Vec<QuantifiedWhereClause<TF>>>)

A "dyn" type is a trait object type created via the "dyn Trait" syntax. In the chalk parser, the traits that the object represents is parsed as a QuantifiedInlineBound, and is then changed to a list of where clauses during lowering.

See the Opaque variant for a discussion about the use of binders here.

Opaque(Binders<Vec<QuantifiedWhereClause<TF>>>)

An "opaque" type is one that is created via the "impl Trait" syntax. They are named so because the concrete type implementing the trait is unknown, and hence the type is opaque to us. The only information that we know of is that this type implements the traits listed by the user.

The "binder" here represents the unknown self type. So, a type like impl for<'a> Fn(&'a u32) would be represented with two-levels of binder, as "depicted" here:

exists<type> {
   vec![
       // A QuantifiedWhereClause:
       forall<region> { ^1: Fn(&^0 u32) }
   ]
}

The outer exists<type> binder indicates that there exists some type that meets the criteria within, but that type is not known. It is referenced within the type using ^1, indicating a bound type with debruijn index 1 (i.e., skipping through one level of binder).

Projection(ProjectionTy<TF>)

A "projection" type corresponds to an (unnormalized) projection like <P0 as Trait<P1..Pn>>::Foo. Note that the trait and all its parameters are fully known.

ForAll(Box<QuantifiedTy<TF>>)

A "higher-ranked" type. In the Rust surface syntax, this can only be a function type (e.g., for<'a> fn(&'a u32)) or a dyn type (e.g., dyn for<'a> SomeTrait<&'a u32>). However, in Chalk's representation, we separate out the for<'a> part from the underlying type, so technically we can represent things like for<'a> SomeStruct<'a>, although that has no meaning in Rust.

BoundVar(usize)

References the binding at the given depth. The index is a de Bruijn index, so it counts back through the in-scope binders, with 0 being the innermost binder. This is used in impls and the like. For example, if we had a rule like for<T> { (T: Clone) :- (T: Copy) }, then T would be represented as a BoundVar(0) (as the for is the innermost binder).

InferenceVar(InferenceVar)

Inference variable defined in the current inference context.

Methods

impl<TF: TypeFamily> Ty<TF>

fn intern(self) -> TF::Type

pub fn bound(&self) -> Option<usize>

If this is a Ty::BoundVar(d), returns Some(d) else None.

pub fn inference_var(&self) -> Option<InferenceVar>

If this is a Ty::InferenceVar(d), returns Some(d) else None.

pub fn is_projection(&self) -> bool

pub fn needs_shift(&self) -> bool

True if this type contains "bound" types/lifetimes, and hence needs to be shifted across binders. This is a very inefficient check, intended only for debug assertions, because I am lazy.