Trait chalk_ir::family::TypeFamily

pub trait TypeFamily: Debug + Copy + Eq + Ord + Hash {
    type InternedType: Debug + Clone + Eq + Ord + Hash;
    type InternedLifetime: Debug + Clone + Eq + Ord + Hash;
    type InternedParameter: Debug + Clone + Eq + Ord + Hash;
    type InternedGoal: Debug + Clone + Eq + Ord + Hash;
    type InternedGoals: Debug + Clone + Eq + Ord + Hash;
    type InternedSubstitution: Debug + Clone + Eq + Ord + Hash;
    type DefId: Debug + Copy + Eq + Ord + Hash;
    fn debug_struct_id(
        struct_id: StructId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result>;

    fn debug_trait_id(
        trait_id: TraitId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result>;

    fn debug_assoc_type_id(
        type_id: AssocTypeId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result>;

    fn debug_alias(alias: &AliasTy<Self>, fmt: &mut Formatter) -> Option<Result>;

    fn intern_ty(ty: TyData<Self>) -> Self::InternedType;

    fn ty_data(ty: &Self::InternedType) -> &TyData<Self>;

    fn intern_lifetime(lifetime: LifetimeData<Self>) -> Self::InternedLifetime;

    fn lifetime_data(lifetime: &Self::InternedLifetime) -> &LifetimeData<Self>;

    fn intern_parameter(data: ParameterData<Self>) -> Self::InternedParameter;

    fn parameter_data(
        lifetime: &Self::InternedParameter
    ) -> &ParameterData<Self>;

    fn intern_goal(data: GoalData<Self>) -> Self::InternedGoal;

    fn goal_data(goal: &Self::InternedGoal) -> &GoalData<Self>;

    fn intern_goals(
        data: impl IntoIterator<Item = Goal<Self>>
    ) -> Self::InternedGoals;

    fn goals_data(goals: &Self::InternedGoals) -> &[Goal<Self>];

    fn intern_substitution<E>(
        data: impl IntoIterator<Item = Result<Parameter<Self>, E>>
    ) -> Result<Self::InternedSubstitution, E>;

    fn substitution_data(
        substitution: &Self::InternedSubstitution
    ) -> &[Parameter<Self>];
}

A "type family" encapsulates the concrete representation of certain "core types" from chalk-ir. All the types in chalk-ir are parameterized by a TF: TypeFamily, and so (e.g.) if they want to store a type, they don't store a Ty<TF> instance directly, but rather prefer a Ty<TF>. You can think of TF::Type as the interned representation (and, indeed, it may well be an interned pointer, e.g. in rustc).

Type families allow chalk to be embedded in different contexts where the concrete representation of core types varies. They also allow us to write generic code that reasons about multiple distinct sets of types by using distinct generic type parameters (e.g., SourceTF and TargetTF) -- even if those type parameters wind up being mapped to the same underlying type families in the end.

Associated Types

type InternedType: Debug + Clone + Eq + Ord + Hash

"Interned" representation of types. In normal user code, Self::InternedType is not referenced Instead, we refer to Ty<Self>, which wraps this type.

An InternedType must be something that can be created from a TyData (by the [intern_ty] method) and then later converted back (by the [ty_data] method). The interned form must also introduce indirection, either via a Box, &, or other pointer type.

type InternedLifetime: Debug + Clone + Eq + Ord + Hash

"Interned" representation of lifetimes. In normal user code, Self::InternedLifetime is not referenced Instead, we refer to Lifetime<Self>, which wraps this type.

An InternedLifetime must be something that can be created from a LifetimeData (by the [intern_lifetime] method) and then later converted back (by the [lifetime_data] method).

type InternedParameter: Debug + Clone + Eq + Ord + Hash

"Interned" representation of a "generic parameter", which can be either a type or a lifetime. In normal user code, Self::InternedParameter is not referenced. Instead, we refer to Parameter<Self>, which wraps this type.

An InternedType is created by intern_parameter and can be converted back to its underlying data via parameter_data.

type InternedGoal: Debug + Clone + Eq + Ord + Hash

"Interned" representation of a "goal". In normal user code, Self::InternedGoal is not referenced. Instead, we refer to Goal<Self>, which wraps this type.

An InternedGoal is created by intern_goal and can be converted back to its underlying data via goal_data.

type InternedGoals: Debug + Clone + Eq + Ord + Hash

"Interned" representation of a list of goals. In normal user code, Self::InternedGoals is not referenced. Instead, we refer to Goals<Self>, which wraps this type.

An InternedGoals is created by intern_goals and can be converted back to its underlying data via goals_data.

type InternedSubstitution: Debug + Clone + Eq + Ord + Hash

"Interned" representation of a "substitution". In normal user code, Self::InternedSubstitution is not referenced. Instead, we refer to Substitution<Self>, which wraps this type.

An InternedSubstitution is created by intern_substitution and can be converted back to its underlying data via substitution_data.

type DefId: Debug + Copy + Eq + Ord + Hash

The core "id" type used for struct-ids and the like.

Loading content...
Required methods

fn debug_struct_id(
struct_id: StructId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a type-kind-id. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific type-family (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_trait_id(
trait_id: TraitId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a type-kind-id. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific type-family (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_assoc_type_id(
type_id: AssocTypeId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a type-kind-id. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific type-family (and hence fully known types).

fn debug_alias(alias: &AliasTy<Self>, fmt: &mut Formatter) -> Option<Result>

Prints the debug representation of an alias. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific type-family (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn intern_ty(ty: TyData<Self>) -> Self::InternedType

Create an "interned" type from ty. This is not normally invoked directly; instead, you invoke TyData::intern (which will ultimately call this method).

fn ty_data(ty: &Self::InternedType) -> &TyData<Self>

Lookup the TyData from an interned type.

fn intern_lifetime(lifetime: LifetimeData<Self>) -> Self::InternedLifetime

Create an "interned" lifetime from lifetime. This is not normally invoked directly; instead, you invoke LifetimeData::intern (which will ultimately call this method).

fn lifetime_data(lifetime: &Self::InternedLifetime) -> &LifetimeData<Self>

Lookup the LifetimeData that was interned to create a InternedLifetime.

fn intern_parameter(data: ParameterData<Self>) -> Self::InternedParameter

Create an "interned" parameter from data. This is not normally invoked directly; instead, you invoke ParameterData::intern (which will ultimately call this method).

fn parameter_data(lifetime: &Self::InternedParameter) -> &ParameterData<Self>

Lookup the LifetimeData that was interned to create a InternedLifetime.

fn intern_goal(data: GoalData<Self>) -> Self::InternedGoal

Create an "interned" goal from data. This is not normally invoked directly; instead, you invoke GoalData::intern (which will ultimately call this method).

fn goal_data(goal: &Self::InternedGoal) -> &GoalData<Self>

Lookup the GoalData that was interned to create a InternedGoal.

fn intern_goals(
data: impl IntoIterator<Item = Goal<Self>>
) -> Self::InternedGoals

Create an "interned" goals from data. This is not normally invoked directly; instead, you invoke GoalsData::intern (which will ultimately call this method).

fn goals_data(goals: &Self::InternedGoals) -> &[Goal<Self>]

Lookup the GoalsData that was interned to create a InternedGoals.

fn intern_substitution<E>(
data: impl IntoIterator<Item = Result<Parameter<Self>, E>>
) -> Result<Self::InternedSubstitution, E>

Create an "interned" substitution from data. This is not normally invoked directly; instead, you invoke SubstitutionData::intern (which will ultimately call this method).

fn substitution_data(
substitution: &Self::InternedSubstitution
) -> &[Parameter<Self>]

Lookup the SubstitutionData that was interned to create a InternedSubstitution.

Loading content...