Module chalk_ir::fold::shift

Shifting of debruijn indices

Structs

DownShifter
A shifter that reduces debruijn indices -- in other words, which lifts a value out from binders. Consider this example:

Shifter
A folder that adjusts debruijn indices by a certain amount.

Traits

Shift
Methods for converting debruijn indices to move values into or out of binders.