Struct chalk_ir::AnswerSubst

pub struct AnswerSubst<I: Interner> {
    pub subst: Substitution<I>,
    pub constraints: Constraints<I>,
    pub delayed_subgoals: Vec<InEnvironment<Goal<I>>>,
}

The resulting substitution after solving a goal.

Fields
subst: Substitution<I>

The substitution result.

NB: The is_trivial routine relies on the fact that subst is folded first.

constraints: Constraints<I>

List of constraints that are part of the answer.

delayed_subgoals: Vec<InEnvironment<Goal<I>>>

Delayed subgoals, used when the solver answered with an (incomplete) Answer (instead of a CompleteAnswer).