Struct chalk_ir::Binders

pub struct Binders<T: HasInterner> {
    pub binders: VariableKinds<T::Interner>,
    pub(crate) value: T,
}

Indicates that the value is universally quantified over N parameters of the given kinds, where N == self.binders.len(). A variable with depth i < N refers to the value at self.binders[i]. Variables with depth >= N are free.

(IOW, we use deBruijn indices, where binders are introduced in reverse order of self.binders.)

Fields
binders: VariableKinds<T::Interner>

The binders that quantify over the value.

value: T

The value being quantified over.

Implementations

impl<I: Interner> Binders<WhereClause<I>>

pub fn into_well_formed_goal(self, interner: I) -> Binders<DomainGoal<I>>

As with WhereClause::into_well_formed_goal, but for a quantified where clause. For example, forall<T> { Implemented(T: Trait)} would map to forall<T> { WellFormed(T: Trait) }.

pub fn into_from_env_goal(self, interner: I) -> Binders<DomainGoal<I>>

As with WhereClause::into_from_env_goal, but mapped over any binders. For example, forall<T> { Implemented(T: Trait)} would map to forall<T> { FromEnv(T: Trait) }.

pub fn trait_id(&self) -> Option<TraitId<I>>

If the underlying where clause is a TraitRef, returns its trait id.

impl<T: Clone + HasInterner> Binders<&T>

pub fn cloned(self) -> Binders<T>

Converts a Binders<&T> to a Binders<T> by cloning T.

impl<T: HasInterner> Binders<T>

pub fn new(binders: VariableKinds<T::Interner>, value: T) -> Self

Create new binders.

pub fn empty(interner: T::Interner, value: T) -> Self

Wraps the given value in a binder without variables, i.e. for<> (value). Since our deBruijn indices count binders, not variables, this is sometimes useful.

pub fn skip_binders(&self) -> &T

Skips the binder and returns the “bound” value. This is a risky thing to do because it’s easy to get confused about De Bruijn indices and the like. skip_binder is only valid when you are either extracting data that has nothing to do with bound vars, or you are being very careful about your depth accounting.

Some examples where skip_binder is reasonable:

extracting the TraitId from a TraitRef;

checking if there are any fields in a StructDatum

pub fn into_value_and_skipped_binders(self) -> (T, VariableKinds<T::Interner>)

Skips the binder and returns the “bound” value as well as the skipped free variables. This is just as risky as skip_binders.

pub fn as_ref(&self) -> Binders<&T>

Converts &Binders<T> to Binders<&T>. Produces new Binders with cloned quantifiers containing a reference to the original value, leaving the original in place.

pub fn map<U, OP>(self, op: OP) -> Binders<U> where
OP: FnOnce(T) -> U,
U: HasInterner<Interner = T::Interner>,

Maps the binders by applying a function.

pub fn filter_map<U, OP>(self, op: OP) -> Option<Binders<U>> where
OP: FnOnce(T) -> Option<U>,
U: HasInterner<Interner = T::Interner>,

Transforms the inner value according to the given function; returns None if the function returns None.

pub fn map_ref<'a, U, OP>(&'a self, op: OP) -> Binders<U> where
OP: FnOnce(&'a T) -> U,
U: HasInterner<Interner = T::Interner>,

Maps a function taking Binders<&T> over &Binders<T>.

pub fn identity_substitution(
&self,
interner: T::Interner
) -> Substitution<T::Interner>

Creates a Substitution containing bound vars such that applying this substitution will not change the value, i.e. ^0.0, ^0.1, ^0.2 and so on.

pub fn with_fresh_type_var(
interner: T::Interner,
op: impl FnOnce(Ty<T::Interner>) -> T
) -> Binders<T>

Creates a fresh binders that contains a single type variable. The result of the closure will be embedded in this binder. Note that you should be careful with what you return from the closure to account for the binder that will be added.

XXX FIXME – this is potentially a pretty footgun-y function.

pub fn len(&self, interner: T::Interner) -> usize

Returns the number of binders.

impl<T, I> Binders<Binders<T>> where
T: Fold<I> + HasInterner<Interner = I>,
T::Result: HasInterner<Interner = I>,
I: Interner,

pub fn fuse_binders(self, interner: T::Interner) -> Binders<T::Result>

This turns two levels of binders (for<A> for<B>) into one level (for<A, B>).

impl<T, I> Binders<T> where
T: Fold<I> + HasInterner<Interner = I>,
I: Interner,

pub fn substitute(
self,
interner: I,
parameters: &impl AsParameters<I> + ?Sized
) -> T::Result

Substitute parameters for the variables introduced by these binders. So if the binders represent (e.g.) <X, Y> { T } and parameters is the slice [A, B], then returns [X => A, Y => B] T.