Struct chalk_ir::Environment

pub struct Environment<I: Interner> {
    pub clauses: ProgramClauses<I>,
}

The set of assumptions we've made so far, and the current number of universal (forall) quantifiers we're within.

Fields
clauses: ProgramClauses<I>

The clauses in the environment.

Implementations

impl<I: Interner> Environment<I>

pub fn new(interner: &I) -> Self

Creates a new environment.

pub fn add_clauses<II>(&self, interner: &I, clauses: II) -> Self where
II: IntoIterator<Item = ProgramClause<I>>,

Adds (an iterator of) clauses to the environment.

pub fn has_compatible_clause(&self, interner: &I) -> bool

True if any of the clauses in the environment have a consequence of Compatible. Panics if the conditions or constraints of that clause are not empty.