Struct chalk_ir::Substitution

pub struct Substitution<I: Interner> {
    interned: I::InternedSubstitution,
}

List of interned elements.

Fields
interned: I::InternedSubstitution
Implementations

impl<I: Interner> Substitution<I>

pub fn debug<'a>(&'a self, interner: &'a I) -> SubstitutionDebug<'a, I>

Show debug output for the substitution.

impl<I: Interner> Substitution<I>

pub fn with_angle(&self, interner: &I) -> Angle<'_, GenericArg<I>>

Displays the substitution in the form < P0, .. Pn >, or (if the substitution is empty) as an empty string.

impl<I: Interner> Substitution<I>

pub fn is_identity_subst(&self, interner: &I) -> bool

A substitution is an identity substitution if it looks like this

?0 := ?0
?1 := ?1
?2 := ?2
...

Basically, each value is mapped to a type or lifetime with its same index.

pub fn apply<T>(&self, value: T, interner: &I) -> T::Result where
T: Fold<I>,

Apply the substitution to a value.

pub fn type_parameters<'a>(
&'a self,
interner: &'a I
) -> impl Iterator<Item = Ty<I>> + 'a

Gets an iterator of all type parameters.

pub(crate) fn compute_flags(&self, interner: &I) -> TypeFlags

Compute type flags for Substitution

impl<I: Interner> Substitution<I>

pub fn interned(&self) -> &I::InternedSubstitution

Get the interned elements.

pub fn as_slice(&self, interner: &I) -> &[GenericArg<I>]

Returns a slice containing the elements.

pub fn at(&self, interner: &I, index: usize) -> &GenericArg<I>

Index into the sequence.

pub fn empty(interner: &I) -> Self

Create an empty sequence.

pub fn is_empty(&self, interner: &I) -> bool

Check whether this is an empty sequence.

pub fn iter(&self, interner: &I) -> Iter<'_, GenericArg<I>>

Get an iterator over the elements of the sequence.

pub fn len(&self, interner: &I) -> usize

Get the length of the sequence.

impl<I: Interner> Substitution<I>

pub fn from_fallible<E>(
interner: &I,
elements: impl IntoIterator<Item = Result<impl CastTo<GenericArg<I>>, E>>
) -> Result<Self, E>

Tries to create a sequence using an iterator of element-like things.

pub fn from_iter(
interner: &I,
elements: impl IntoIterator<Item = impl CastTo<GenericArg<I>>>
) -> Self

Create a sequence from elements

pub fn from1(interner: &I, element: impl CastTo<GenericArg<I>>) -> Self

Create a sequence from a single element.