Struct chalk_ir::Ty

pub struct Ty<I: Interner> {
    pub(crate) interned: I::InternedType,
}

A Rust type. The actual type data is stored in TyKind.

Fields
interned: I::InternedType
Implementations

impl<I: Interner> Ty<I>

pub fn new(interner: I, data: impl CastTo<TyKind<I>>) -> Self

Creates a type from TyKind.

pub fn interned(&self) -> &I::InternedType

Gets the interned type.

pub fn data(&self, interner: I) -> &TyData<I>

Gets the underlying type data.

pub fn kind(&self, interner: I) -> &TyKind<I>

Gets the underlying type kind.

pub fn from_env(&self) -> FromEnv<I>

Creates a FromEnv constraint using this type.

pub fn well_formed(&self) -> WellFormed<I>

Creates a WF-constraint for this type.

pub fn into_from_env_goal(self, interner: I) -> DomainGoal<I>

Creates a domain goal FromEnv(T) where T is this type.

pub fn bound_var(&self, interner: I) -> Option<BoundVar>

If this is a TyKind::BoundVar(d), returns Some(d) else None.

pub fn inference_var(&self, interner: I) -> Option<InferenceVar>

If this is a TyKind::InferenceVar(d), returns Some(d) else None.

pub fn is_general_var( &self, interner: I, binders: &CanonicalVarKinds<I> ) -> bool

Returns true if this is a BoundVar or an InferenceVar of TyVariableKind::General.

pub fn is_alias(&self, interner: I) -> bool

Returns true if this is an Alias.

pub fn is_integer(&self, interner: I) -> bool

Returns true if this is an IntTy or UintTy.

pub fn is_float(&self, interner: I) -> bool

Returns true if this is a FloatTy.

pub fn adt_id(&self, interner: I) -> Option<AdtId<I>>

Returns Some(adt_id) if this is an ADT, None otherwise

pub fn needs_shift(&self, interner: I) -> bool

True if this type contains “bound” types/lifetimes, and hence needs to be shifted across binders. This is a very inefficient check, intended only for debug assertions, because I am lazy.