Trait chalk_ir::Substitute

pub trait Substitute<I: Interner>: AsParameters<I> {
    fn apply<T: Fold<I>>(&self, value: T, interner: I) -> T::Result;
}

An extension trait to anything that can be represented as list of GenericArgs that signifies that it can applied as a substituion to a value

Required methods

fn apply<T: Fold<I>>(&self, value: T, interner: I) -> T::Result

Apply the substitution to a value.