Trait chalk_ir::ToGenericArg

pub trait ToGenericArg<I: Interner> {
    fn to_generic_arg_at_depth(
        &self, 
        interner: I, 
        debruijn: DebruijnIndex
    ) -> GenericArg<I>;

    fn to_generic_arg(&self, interner: I) -> GenericArg<I> { ... }
}

Utility for converting a list of all the binders into scope into references to those binders. Simply pair the binders with the indices, and invoke to_generic_arg() on the (binder, index) pair. The result will be a reference to a bound variable of appropriate kind at the corresponding index.

Required methods

fn to_generic_arg_at_depth(
&self,
interner: I,
debruijn: DebruijnIndex
) -> GenericArg<I>

Converts the binders at the specified depth to references to those binders.

Provided methods

fn to_generic_arg(&self, interner: I) -> GenericArg<I>

Converts the binders in scope to references to those binders.