Trait chalk_ir::UnificationDatabase

pub trait UnificationDatabase<I>
where
    Self: Debug,
    I: Interner,
{
    // Required methods
    fn fn_def_variance(&self, fn_def_id: FnDefId<I>) -> Variances<I>;
    fn adt_variance(&self, adt_id: AdtId<I>) -> Variances<I>;
}

Logic to decide the Variance for a given subst

Required Methods

fn fn_def_variance(&self, fn_def_id: FnDefId<I>) -> Variances<I>

Gets the variances for the substitution of a fn def

fn adt_variance(&self, adt_id: AdtId<I>) -> Variances<I>

Gets the variances for the substitution of a adt