Module chalk_ir::visit

Traits for visiting bits of IR.

Re-exports

pub use visitors::VisitExt;

Modules

binder_impls 🔒

This module contains impls of TypeVisitable for those types that introduce binders.

boring_impls 🔒

This module contains “rote and uninteresting” impls of TypeVisitable for various types. In general, we prefer to derive TypeVisitable, but sometimes that doesn’t work for whatever reason.

visitors

TypeVisitor helpers

Traits

TypeSuperVisitable

For types where “visit” invokes a callback on the visitor, the TypeSuperVisitable trait captures the recursive behavior that visits all the contents of the type.

TypeVisitable

Applies the given visitor to a value, producing a visited result of type TypeVisitor::Result.

TypeVisitor

A “visitor” recursively folds some term – that is, some bit of IR, such as a Goal, and computes a value as a result.