Module chalk_ir::debug

Debug impls for types.

Structs

Angle

Wraps debug output in angle brackets (<>).

CanonicalDisplay

Helper struct for displaying canonicalized items.

Fmt

Wrapper to allow forwarding to Display::fmt, Debug::fmt, etc.

GenericArgDataInnerDebug

Helper struct for showing debug output for GenericArgData.

GoalsDebug

Helper struct for showing debug output for Goals.

OpaqueTyDebug

Helper struct for showing debug output for opaque types.

ProgramClauseImplicationDebug

Helper struct for showing debug output for program clause implications.

ProjectionTyDebug

Helper struct for showing debug output for projection types.

SeparatorTraitRef

Trait ref with associated separator used for debug output.

SeparatorTraitRefDebug

Helper struct for showing debug output for the SeperatorTraitRef.

SubstitutionDebug

Helper struct for showing debug output for substitutions.

TyKindDebug

Helper struct for showing debug output for application types.

VariableKindsDebug 🔒

VariableKindsInnerDebug

Helper struct for showing debug output for VariableKinds.