Enum chalk_ir::TyKind

pub enum TyKind<I: Interner> {

    Adt(AdtId<I>, Substitution<I>),
    AssociatedType(AssocTypeId<I>, Substitution<I>),
    Scalar(Scalar),
    Tuple(usize, Substitution<I>),
    Array(Ty<I>, Const<I>),
    Slice(Ty<I>),
    Raw(Mutability, Ty<I>),
    Ref(Mutability, Lifetime<I>, Ty<I>),
    OpaqueType(OpaqueTyId<I>, Substitution<I>),
    FnDef(FnDefId<I>, Substitution<I>),
    Str,
    Never,
    Closure(ClosureId<I>, Substitution<I>),
    Coroutine(CoroutineId<I>, Substitution<I>),
    CoroutineWitness(CoroutineId<I>, Substitution<I>),
    Foreign(ForeignDefId<I>),
    Error,
    Placeholder(PlaceholderIndex),
    Dyn(DynTy<I>),
    Alias(AliasTy<I>),
    Function(FnPointer<I>),
    BoundVar(BoundVar),
    InferenceVar(InferenceVar, TyVariableKind),

}

Type data, which holds the actual type information.

Variants

Adt(AdtId<I>, Substitution<I>)

Abstract data types, i.e., structs, unions, or enumerations. For example, a type like Vec<T>.

AssociatedType(AssocTypeId<I>, Substitution<I>)

an associated type like Iterator::Item; see AssociatedType for details

Scalar(Scalar)

a scalar type like bool or u32

Tuple(usize, Substitution<I>)

a tuple of the given arity

Array(Ty<I>, Const<I>)

an array type like [T; N]

Slice(Ty<I>)

a slice type like [T]

Raw(Mutability, Ty<I>)

a raw pointer type like *const T or *mut T

Ref(Mutability, Lifetime<I>, Ty<I>)

a reference type like &T or &mut T

OpaqueType(OpaqueTyId<I>, Substitution<I>)

a placeholder for opaque types like impl Trait

FnDef(FnDefId<I>, Substitution<I>)

a function definition

Str

the string primitive type

Never

the never type !

Closure(ClosureId<I>, Substitution<I>)

A closure.

Coroutine(CoroutineId<I>, Substitution<I>)

A coroutine.

CoroutineWitness(CoroutineId<I>, Substitution<I>)

A coroutine witness.

Foreign(ForeignDefId<I>)

foreign types

Error

This can be used to represent an error, e.g. during name resolution of a type. Chalk itself will not produce this, just pass it through when given.

Placeholder(PlaceholderIndex)

instantiated from a universally quantified type, e.g., from forall<T> { .. }. Stands in as a representative of “some unknown type”.

Dyn(DynTy<I>)

A “dyn” type is a trait object type created via the “dyn Trait” syntax. In the chalk parser, the traits that the object represents is parsed as a QuantifiedInlineBound, and is then changed to a list of where clauses during lowering.

See the Opaque variant for a discussion about the use of binders here.

Alias(AliasTy<I>)

An “alias” type represents some form of type alias, such as:

An associated type projection like <T as Iterator>::Item

impl Trait types

Named type aliases like type Foo<X> = Vec<X>

Function(FnPointer<I>)

A function type such as for<'a> fn(&'a u32). Note that “higher-ranked” types (starting with for<>) are either function types or dyn types, and do not appear otherwise in Rust surface syntax.

BoundVar(BoundVar)

References the binding at the given depth. The index is a [de Bruijn index], so it counts back through the in-scope binders.

InferenceVar(InferenceVar, TyVariableKind)

Inference variable defined in the current inference context.

Implementations

impl<I: Interner> TyKind<I>

pub fn debug(&self, interner: I) -> TyKindDebug<'_, I>

Show debug output for the application type.

impl<I: Interner> TyKind<I>

pub fn intern(self, interner: I) -> Ty<I>

Casts the type data to a type.

pub fn compute_flags(&self, interner: I) -> TypeFlags

Compute type flags for a TyKind