Trait chalk_ir::family::HasTypeFamily

pub trait HasTypeFamily {
    type TypeFamily: TypeFamily;
}

Implemented by types that have an associated type family (which are virtually all of the types in chalk-ir, for example). This lets us map from a type like Ty<TF> to the parameter TF.

It's particularly useful for writing Fold impls for generic types like Binder<T>, since it allows us to figure out the type family of T.

Associated Types

type TypeFamily: TypeFamily

Loading content...