Trait chalk_ir::family::TypeFamily

pub trait TypeFamily: Debug + Copy + Eq + Ord + Hash {
    type InternedType: Debug + Clone + Eq + Ord + Hash;
    type InternedLifetime: Debug + Clone + Eq + Ord + Hash;
    fn debug_projection(
        projection: &ProjectionTy<Self>, 
        fmt: &mut Formatter
    ) -> Result;

    fn intern_ty(ty: TyData<Self>) -> Self::InternedType;

    fn ty_data(ty: &Self::InternedType) -> &TyData<Self>;

    fn intern_lifetime(lifetime: LifetimeData<Self>) -> Self::InternedLifetime;

    fn lifetime_data(lifetime: &Self::InternedLifetime) -> &LifetimeData<Self>;
}

A "type family" encapsulates the concrete representation of certain "core types" from chalk-ir. All the types in chalk-ir are parameterized by a TF: TypeFamily, and so (e.g.) if they want to store a type, they don't store a Ty<TF> instance directly, but rather prefer a Ty<TF>. You can think of TF::Type as the interned representation (and, indeed, it may well be an interned pointer, e.g. in rustc).

Type families allow chalk to be embedded in different contexts where the concrete representation of core types varies. They also allow us to write generic code that reasons about multiple distinct sets of types by using distinct generic type parameters (e.g., SourceTF and TargetTF) -- even if those type parameters wind up being mapped to the same underlying type families in the end.

Associated Types

type InternedType: Debug + Clone + Eq + Ord + Hash

"Interned" representation of types. You can use the Lookup trait to convert this to a Ty<Self>.

type InternedLifetime: Debug + Clone + Eq + Ord + Hash

"Interned" representation of lifetimes. You can use the Lookup trait to convert this to a Lifetime<Self>.

Loading content...
Required methods

fn debug_projection(
projection: &ProjectionTy<Self>,
fmt: &mut Formatter
) -> Result

Prints the debug representation of a projection. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific type-family (and hence fully known types).

fn intern_ty(ty: TyData<Self>) -> Self::InternedType

Create an "interned" type from ty. This is not normally invoked directly; instead, you invoke TyData::intern (which will ultimately call this method).

fn ty_data(ty: &Self::InternedType) -> &TyData<Self>

Lookup the TyData from an interned type.

fn intern_lifetime(lifetime: LifetimeData<Self>) -> Self::InternedLifetime

Create an "interned" lifetime from lifetime. This is not normally invoked directly; instead, you invoke LifetimeData::intern (which will ultimately call this method).

fn lifetime_data(lifetime: &Self::InternedLifetime) -> &LifetimeData<Self>

Loading content...