Module chalk_ir::fold

Traits for transforming bits of IR.

Modules

shift

subst

Structs

Subst

Traits

DefaultFreeVarFolder
A convenience trait. If you implement this, you get an implementation of FreeVarFolder for free that simply ignores free values (that is, it replaces them with themselves).

DefaultInferenceFolder
A convenience trait. If you implement this, you get an implementation of InferenceFolder for free that simply ignores inference values (that is, it replaces them with themselves).

DefaultPlaceholderFolder
A convenience trait. If you implement this, you get an implementation of PlaceholderFolder for free that simply ignores placeholder values (that is, it replaces them with themselves).

DefaultTypeFolder
A convenience trait that indicates that this folder doesn't take any action on types in particular, but just recursively folds their contents (note that free variables that are encountered in that process may still be substituted). The vast majority of folders implement this trait.

Fold
Applies the given folder to a value.

Folder
A "folder" is a transformer that can be used to make a copy of some term -- that is, some bit of IR, such as a Goal -- with certain changes applied. The idea is that it contains methods that let you swap types/lifetimes for new types/lifetimes; meanwhile, each bit of IR implements the Fold trait which, given a Folder, will reconstruct itself, invoking the folder's methods to transform each of the types/lifetimes embedded within.

FreeVarFolder
The methods for folding free variables. These are BoundVar instances where the binder is not something we folded over. This is used when you are instanting previously bound things with some replacement.

InferenceFolder

PlaceholderFolder

TypeFolder

Functions

super_fold_lifetime

super_fold_ty