Module chalk_ir::fold::in_place

Subroutines to help implementers of Fold avoid unnecessary heap allocations.

Structs

VecMappedInPlace
Takes ownership of a Vec that is being mapped in place, cleaning up if the map fails.

Functions

fallible_map_box
Maps a Box<T> to a Box<U>, reusing the underlying storage if possible.

fallible_map_vec
Maps a Vec<T> to a Vec<U>, reusing the underlying storage if possible.

is_layout_identical

is_zst