Trait chalk_ir::fold::shift::Shift

pub trait Shift<TF: TypeFamily>: Fold<TF, TF> {
    fn shifted_in(&self, adjustment: usize) -> Self::Result;

    fn shifted_out(&self, adjustment: usize) -> Fallible<Self::Result>;
}

Methods for converting debruijn indices to move values into or out of binders.

Required methods

fn shifted_in(&self, adjustment: usize) -> Self::Result

Shifts debruijn indices in self up, which is used when a value is being placed under additional levels of binders.

For example, if we had some goal like:

T: Trait<?X>

where ?X refers to some inference variable (and hence has depth 3), we might use up_shift when constructing a goal like:

exists<U> { T = U, T: Trait<?X> }

This is because, internally, the inference variable ?X (as well as the new quantified variable U) are going to be represented by debruijn indices. So if the index of X is zero, then while originally we might have had T: Trait<?0>, inside the exists we want to represent X with ?1, to account for the binder:

    exists { T = ?0, T: Trait<?1> }
                 ^^           ^^ refers to `?X`
                 refers to `U`

fn shifted_out(&self, adjustment: usize) -> Fallible<Self::Result>

Shifts debruijn indices in self down, hence removing a value from binders. This will fail with Err(NoSolution) in the case that the value refers to something from one of those binders.

Consider the final example from up_shift:

    exists { T = ?0, T: Trait<?1> }
                 ^^           ^^ refers to `?X`
                 refers to `U`

If we down_shift the T: Trait<?1> goal by 1, we will get T: Trait<?0>, which is what we started with. In other words, we will have extracted it from the exists binder.

But if we try to down_shift the T = ?0 goal by 1, we will get Err, because it refers to the type bound by the exists.

Loading content...