Trait chalk_ir::fold::Fold

pub trait Fold: Debug {
    type Result: Fold;
    fn fold_with(
        &self, 
        folder: &mut dyn Folder, 
        binders: usize
    ) -> Fallible<Self::Result>;
}

Applies the given folder to a value.

Associated Types

type Result: Fold

The type of value that will be produced once folding is done. Typically this is Self, unless Self contains borrowed values, in which case owned values are produced (for example, one can fold over a &T value where T: Fold, in which case you get back a T, not a &T).

Loading content...
Required methods

fn fold_with(
&self,
folder: &mut dyn Folder,
binders: usize
) -> Fallible<Self::Result>

Apply the given folder folder to self; binders is the number of binders that are in scope when beginning the folder. Typically binders starts as 0, but is adjusted when we encounter Binders<T> in the IR or other similar constructs.

Loading content...