Trait chalk_ir::fold::Folder

pub trait Folder<TF: TypeFamily, TTF: TypeFamily>: FreeVarFolder<TTF> + InferenceFolder<TTF> + PlaceholderFolder<TTF> + TypeFolder<TF, TTF> { }

A "folder" is a transformer that can be used to make a copy of some term -- that is, some bit of IR, such as a Goal -- with certain changes applied. The idea is that it contains methods that let you swap types/lifetimes for new types/lifetimes; meanwhile, each bit of IR implements the Fold trait which, given a Folder, will reconstruct itself, invoking the folder's methods to transform each of the types/lifetimes embedded within.

Type families

The Folder trait has two type parameters, TF and TTF:

TF is the "source type family" that we are folding from

TTF is the "target type family" that we are folding into

Often, both are the same.

Usage patterns

Substituting for free variables

Most of the time, though, we are not interested in adjust arbitrary types/lifetimes, but rather just free variables (even more often, just free existential variables) that appear within the term.

For this reason, the Folder trait extends two other traits that contain methods that are invoked when just those particular

In particular, folders can intercept references to free variables (either existentially or universally quantified) and replace them with other types/lifetimes as appropriate.

To create a folder F, one never implements Folder directly, but instead implements one of each of these three sub-traits:

FreeVarFolder -- folds BoundVar instances that appear free in the term being folded (use DefaultFreeVarFolder to ignore/forbid these altogether)

InferenceFolder -- folds existential InferenceVar instances that appear in the term being folded (use DefaultInferenceFolder to ignore/forbid these altogether)

PlaceholderFolder -- folds universal Placeholder instances that appear in the term being folded (use DefaultPlaceholderFolder to ignore/forbid these altogether)

To apply a folder, use the Fold::fold_with method, like so

let x = x.fold_with(&mut folder, 0);