Trait chalk_ir::fold::Folder

pub trait Folder: FreeVarFolder + InferenceFolder + PlaceholderFolder + TypeFolder {
    fn to_dyn(&mut self) -> &mut dyn Folder;
}

A "folder" is a transformer that can be used to make a copy of some term -- that is, some bit of IR, such as a Goal -- with certain changes applied. The idea is that it contains methods that let you swap types/lifetimes for new types/lifetimes; meanwhile, each bit of IR implements the Fold trait which, given a Folder, will reconstruct itself, invoking the folder's methods to transform each of the types/lifetimes embedded within.

Usage patterns

Substituting for free variables

Most of the time, though, we are not interested in adjust arbitrary types/lifetimes, but rather just free variables (even more often, just free existential variables) that appear within the term.

For this reason, the Folder trait extends two other traits that contain methods that are invoked when just those particular

In particular, folders can intercept references to free variables (either existentially or universally quantified) and replace them with other types/lifetimes as appropriate.

To create a folder F, one never implements Folder directly, but instead implements one of each of these three sub-traits:

FreeVarFolder -- folds BoundVar instances that appear free in the term being folded (use DefaultFreeVarFolder to ignore/forbid these altogether)

InferenceFolder -- folds existential InferenceVar instances that appear in the term being folded (use DefaultInferenceFolder to ignore/forbid these altogether)

PlaceholderFolder -- folds universal Placeholder instances that appear in the term being folded (use DefaultPlaceholderFolder to ignore/forbid these altogether)

To apply a folder, use the Fold::fold_with method, like so

let x = x.fold_with(&mut folder, 0);

Required methods

fn to_dyn(&mut self) -> &mut dyn Folder

Returns a "dynamic" version of this trait. There is no particular reason to require this, except that I didn't feel like making super_fold_ty generic for no reason.

Loading content...