Trait chalk_ir::fold::Folder

pub trait Folder<TF: TypeFamily, TTF: TargetTypeFamily<TF> = TF> {
    fn as_dyn(&mut self) -> &mut dyn Folder<TF, TTF>;

    fn fold_ty(&mut self, ty: &Ty<TF>, binders: usize) -> Fallible<Ty<TTF>> { ... }

    fn fold_lifetime(
        &mut self, 
        lifetime: &Lifetime<TF>, 
        binders: usize
    ) -> Fallible<Lifetime<TTF>> { ... }

    fn fold_program_clause(
        &mut self, 
        clause: &ProgramClause<TF>, 
        binders: usize
    ) -> Fallible<ProgramClause<TTF>> { ... }

    fn fold_goal(
        &mut self, 
        goal: &Goal<TF>, 
        binders: usize
    ) -> Fallible<Goal<TTF>> { ... }

    fn forbid_free_vars(&self) -> bool { ... }

    fn fold_free_var_ty(
        &mut self, 
        depth: usize, 
        binders: usize
    ) -> Fallible<Ty<TTF>> { ... }

    fn fold_free_var_lifetime(
        &mut self, 
        depth: usize, 
        binders: usize
    ) -> Fallible<Lifetime<TTF>> { ... }

    fn forbid_free_placeholders(&self) -> bool { ... }

    fn fold_free_placeholder_ty(
        &mut self, 
        universe: PlaceholderIndex, 
        _binders: usize
    ) -> Fallible<Ty<TTF>> { ... }

    fn fold_free_placeholder_lifetime(
        &mut self, 
        universe: PlaceholderIndex, 
        _binders: usize
    ) -> Fallible<Lifetime<TTF>> { ... }

    fn forbid_inference_vars(&self) -> bool { ... }

    fn fold_inference_ty(
        &mut self, 
        var: InferenceVar, 
        _binders: usize
    ) -> Fallible<Ty<TTF>> { ... }

    fn fold_inference_lifetime(
        &mut self, 
        var: InferenceVar, 
        _binders: usize
    ) -> Fallible<Lifetime<TTF>> { ... }
}

A "folder" is a transformer that can be used to make a copy of some term -- that is, some bit of IR, such as a Goal -- with certain changes applied. The idea is that it contains methods that let you swap types/lifetimes for new types/lifetimes; meanwhile, each bit of IR implements the Fold trait which, given a Folder, will reconstruct itself, invoking the folder's methods to transform each of the types/lifetimes embedded within.

Type families

The Folder trait has two type parameters, TF and TTF:

TF is the "source type family" that we are folding from

TTF is the "target type family" that we are folding into

Often, both are the same.

Usage patterns

Substituting for free variables

Most of the time, though, we are not interested in adjust arbitrary types/lifetimes, but rather just free variables (even more often, just free existential variables) that appear within the term.

For this reason, the Folder trait extends two other traits that contain methods that are invoked when just those particular

In particular, folders can intercept references to free variables (either existentially or universally quantified) and replace them with other types/lifetimes as appropriate.

To create a folder F, one never implements Folder directly, but instead implements one of each of these three sub-traits:

FreeVarFolder -- folds BoundVar instances that appear free in the term being folded (use DefaultFreeVarFolder to ignore/forbid these altogether)

InferenceFolder -- folds existential InferenceVar instances that appear in the term being folded (use DefaultInferenceFolder to ignore/forbid these altogether)

PlaceholderFolder -- folds universal Placeholder instances that appear in the term being folded (use DefaultPlaceholderFolder to ignore/forbid these altogether)

To apply a folder, use the Fold::fold_with method, like so

let x = x.fold_with(&mut folder, 0);

Required methods

fn as_dyn(&mut self) -> &mut dyn Folder<TF, TTF>

Creates a dyn value from this folder. Unfortunately, this must be added manually to each impl of Folder; it permits the default implements below to create a &mut dyn Folder from Self without knowing what Self is (by invoking this method). Effectively, this limits impls of Folder to types for which we are able to create a dyn value (i.e., not [T] types).

Loading content...
Provided methods

fn fold_ty(&mut self, ty: &Ty<TF>, binders: usize) -> Fallible<Ty<TTF>>

Top-level callback: invoked for each Ty<TF> that is encountered when folding. By default, invokes super_fold_with, which will in turn invoke the more specialized folding methods below, like fold_free_var_ty.

fn fold_lifetime(
&mut self,
lifetime: &Lifetime<TF>,
binders: usize
) -> Fallible<Lifetime<TTF>>

Top-level callback: invoked for each Lifetime<TF> that is encountered when folding. By default, invokes super_fold_with, which will in turn invoke the more specialized folding methods below, like fold_free_lifetime_ty.

fn fold_program_clause(
&mut self,
clause: &ProgramClause<TF>,
binders: usize
) -> Fallible<ProgramClause<TTF>>

Invoked for every program clause. By default, recursively folds the goals contents.

fn fold_goal(&mut self, goal: &Goal<TF>, binders: usize) -> Fallible<Goal<TTF>>

Invoked for every goal. By default, recursively folds the goals contents.

fn forbid_free_vars(&self) -> bool

If overridden to return true, then folding will panic if a free variable is encountered. This should be done if free type/lifetime variables are not expected.

fn fold_free_var_ty(
&mut self,
depth: usize,
binders: usize
) -> Fallible<Ty<TTF>>

Invoked for TyData::BoundVar instances that are not bound within the type being folded over:

depth is the depth of the TyData::BoundVar; this has been adjusted to account for binders in scope.

binders is the number of binders in scope.

This should return a type suitable for a context with binders in scope.

fn fold_free_var_lifetime(
&mut self,
depth: usize,
binders: usize
) -> Fallible<Lifetime<TTF>>

As fold_free_var_ty, but for lifetimes.

fn forbid_free_placeholders(&self) -> bool

If overriden to return true, we will panic when a free placeholder type/lifetime is encountered.

fn fold_free_placeholder_ty(
&mut self,
universe: PlaceholderIndex,
_binders: usize
) -> Fallible<Ty<TTF>>

Invoked for each occurrence of a placeholder type; these are used when we instantiate binders universally. Returns a type to use instead, which should be suitably shifted to account for binders.

universe is the universe of the TypeName::ForAll that was found

binders is the number of binders in scope

fn fold_free_placeholder_lifetime(
&mut self,
universe: PlaceholderIndex,
_binders: usize
) -> Fallible<Lifetime<TTF>>

As with fold_free_placeholder_ty, but for lifetimes.

fn forbid_inference_vars(&self) -> bool

If overridden to return true, inference variables will trigger panics when folded. Used when inference variables are unexpected.

fn fold_inference_ty(
&mut self,
var: InferenceVar,
_binders: usize
) -> Fallible<Ty<TTF>>

Invoked for each occurrence of a inference type; these are used when we instantiate binders universally. Returns a type to use instead, which should be suitably shifted to account for binders.

universe is the universe of the TypeName::ForAll that was found

binders is the number of binders in scope

fn fold_inference_lifetime(
&mut self,
var: InferenceVar,
_binders: usize
) -> Fallible<Lifetime<TTF>>

As with fold_free_inference_ty, but for lifetimes.

Loading content...