Trait chalk_ir::fold::FreeVarFolder

pub trait FreeVarFolder {
    fn fold_free_var_ty(&mut self, depth: usize, binders: usize) -> Fallible<Ty>;

    fn fold_free_var_lifetime(
        &mut self, 
        depth: usize, 
        binders: usize
    ) -> Fallible<Lifetime>;
}

The methods for folding free variables. These are BoundVar instances where the binder is not something we folded over. This is used when you are instanting previously bound things with some replacement.

Required methods

fn fold_free_var_ty(&mut self, depth: usize, binders: usize) -> Fallible<Ty>

Invoked for Ty::BoundVar instances that are not bound within the type being folded over:

depth is the depth of the Ty::BoundVar; this has been adjusted to account for binders in scope.

binders is the number of binders in scope.

This should return a type suitable for a context with binders in scope.

fn fold_free_var_lifetime(
&mut self,
depth: usize,
binders: usize
) -> Fallible<Lifetime>

As fold_free_var_ty, but for lifetimes.

Loading content...