Trait chalk_ir::interner::HasInterner

pub trait HasInterner {
    type Interner: Interner;
}

Implemented by types that have an associated interner (which are virtually all of the types in chalk-ir, for example). This lets us map from a type like Ty<I> to the parameter I.

It’s particularly useful for writing Fold impls for generic types like Binder<T>, since it allows us to figure out the interner of T.

Associated Types

type Interner: Interner

The interner associated with the type.