Trait chalk_ir::interner::Interner

pub trait Interner: Debug + Copy + Eq + Ord + Hash {
    type InternedType: Debug + Clone + Eq + Hash;
    type InternedLifetime: Debug + Clone + Eq + Hash;
    type InternedParameter: Debug + Clone + Eq + Hash;
    type InternedGoal: Debug + Clone + Eq + Hash;
    type InternedGoals: Debug + Clone + Eq + Hash;
    type InternedSubstitution: Debug + Clone + Eq + Hash;
    type InternedProgramClauses: Debug + Clone + Eq + Hash;
    type InternedProgramClause: Debug + Clone + Eq + Hash;
    type InternedQuantifiedWhereClauses: Debug + Clone + Eq + Hash;
    type InternedParameterKinds: Debug + Clone + Eq + Hash;
    type InternedCanonicalVarKinds: Debug + Clone + Eq + Hash;
    type DefId: Debug + Copy + Eq + Ord + Hash;
    type Identifier: Debug + Clone + Eq + Hash;
    fn intern_ty(&self, ty: TyData<Self>) -> Self::InternedType;

    fn ty_data<'a>(&self, ty: &'a Self::InternedType) -> &'a TyData<Self>;

    fn intern_lifetime(
        &self, 
        lifetime: LifetimeData<Self>
    ) -> Self::InternedLifetime;

    fn lifetime_data<'a>(
        &self, 
        lifetime: &'a Self::InternedLifetime
    ) -> &'a LifetimeData<Self>;

    fn intern_parameter(
        &self, 
        data: ParameterData<Self>
    ) -> Self::InternedParameter;

    fn parameter_data<'a>(
        &self, 
        lifetime: &'a Self::InternedParameter
    ) -> &'a ParameterData<Self>;

    fn intern_goal(&self, data: GoalData<Self>) -> Self::InternedGoal;

    fn goal_data<'a>(&self, goal: &'a Self::InternedGoal) -> &'a GoalData<Self>;

    fn intern_goals<E>(
        &self, 
        data: impl IntoIterator<Item = Result<Goal<Self>, E>>
    ) -> Result<Self::InternedGoals, E>;

    fn goals_data<'a>(&self, goals: &'a Self::InternedGoals) -> &'a [Goal<Self>];

    fn intern_substitution<E>(
        &self, 
        data: impl IntoIterator<Item = Result<Parameter<Self>, E>>
    ) -> Result<Self::InternedSubstitution, E>;

    fn substitution_data<'a>(
        &self, 
        substitution: &'a Self::InternedSubstitution
    ) -> &'a [Parameter<Self>];

    fn intern_program_clause(
        &self, 
        data: ProgramClauseData<Self>
    ) -> Self::InternedProgramClause;

    fn program_clause_data<'a>(
        &self, 
        clause: &'a Self::InternedProgramClause
    ) -> &'a ProgramClauseData<Self>;

    fn intern_program_clauses<E>(
        &self, 
        data: impl IntoIterator<Item = Result<ProgramClause<Self>, E>>
    ) -> Result<Self::InternedProgramClauses, E>;

    fn program_clauses_data<'a>(
        &self, 
        clauses: &'a Self::InternedProgramClauses
    ) -> &'a [ProgramClause<Self>];

    fn intern_quantified_where_clauses<E>(
        &self, 
        data: impl IntoIterator<Item = Result<QuantifiedWhereClause<Self>, E>>
    ) -> Result<Self::InternedQuantifiedWhereClauses, E>;

    fn quantified_where_clauses_data<'a>(
        &self, 
        clauses: &'a Self::InternedQuantifiedWhereClauses
    ) -> &'a [QuantifiedWhereClause<Self>];

    fn intern_parameter_kinds<E>(
        &self, 
        data: impl IntoIterator<Item = Result<ParameterKind<()>, E>>
    ) -> Result<Self::InternedParameterKinds, E>;

    fn parameter_kinds_data<'a>(
        &self, 
        parameter_kinds: &'a Self::InternedParameterKinds
    ) -> &'a [ParameterKind<()>];

    fn intern_canonical_var_kinds<E>(
        &self, 
        data: impl IntoIterator<Item = Result<ParameterKind<UniverseIndex>, E>>
    ) -> Result<Self::InternedCanonicalVarKinds, E>;

    fn canonical_var_kinds_data<'a>(
        &self, 
        canonical_var_kinds: &'a Self::InternedCanonicalVarKinds
    ) -> &'a [ParameterKind<UniverseIndex>];

    fn debug_struct_id(
        struct_id: StructId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_trait_id(
        trait_id: TraitId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_assoc_type_id(
        type_id: AssocTypeId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_opaque_ty_id(
        opaque_ty_id: OpaqueTyId<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_alias(alias: &AliasTy<Self>, fmt: &mut Formatter) -> Option<Result> { ... }

    fn debug_projection_ty(
        projection_ty: &ProjectionTy<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_opaque_ty(
        opaque_ty: &OpaqueTy<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_ty(ty: &Ty<Self>, fmt: &mut Formatter) -> Option<Result> { ... }

    fn debug_lifetime(
        lifetime: &Lifetime<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_parameter(
        parameter: &Parameter<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_parameter_kinds(
        parameter_kinds: &ParameterKinds<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_parameter_kinds_with_angles(
        parameter_kinds: &ParameterKinds<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_canonical_var_kinds(
        canonical_var_kinds: &CanonicalVarKinds<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_goal(goal: &Goal<Self>, fmt: &mut Formatter) -> Option<Result> { ... }

    fn debug_goals(goals: &Goals<Self>, fmt: &mut Formatter) -> Option<Result> { ... }

    fn debug_program_clause_implication(
        pci: &ProgramClauseImplication<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_program_clause(
        clause: &ProgramClause<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_program_clauses(
        clauses: &ProgramClauses<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_application_ty(
        application_ty: &ApplicationTy<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_substitution(
        substitution: &Substitution<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_separator_trait_ref(
        separator_trait_ref: &SeparatorTraitRef<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }

    fn debug_quantified_where_clauses(
        clauses: &QuantifiedWhereClauses<Self>, 
        fmt: &mut Formatter
    ) -> Option<Result> { ... }
}

A "interner" encapsulates the concrete representation of certain "core types" from chalk-ir. All the types in chalk-ir are parameterized by a I: Interner, and so (e.g.) if they want to store a type, they don't store a Ty<I> instance directly, but rather prefer a Ty<I>. You can think of I::Type as the interned representation (and, indeed, it may well be an interned pointer, e.g. in rustc).

Type families allow chalk to be embedded in different contexts where the concrete representation of core types varies. They also allow us to write generic code that reasons about multiple distinct sets of types by using distinct generic type parameters (e.g., SourceI and TargetI) -- even if those type parameters wind up being mapped to the same underlying type families in the end.

Associated Types

type InternedType: Debug + Clone + Eq + Hash

"Interned" representation of types. In normal user code, Self::InternedType is not referenced Instead, we refer to Ty<Self>, which wraps this type.

An InternedType must be something that can be created from a TyData (by the [intern_ty] method) and then later converted back (by the [ty_data] method). The interned form must also introduce indirection, either via a Box, &, or other pointer type.

type InternedLifetime: Debug + Clone + Eq + Hash

"Interned" representation of lifetimes. In normal user code, Self::InternedLifetime is not referenced Instead, we refer to Lifetime<Self>, which wraps this type.

An InternedLifetime must be something that can be created from a LifetimeData (by the [intern_lifetime] method) and then later converted back (by the [lifetime_data] method).

type InternedParameter: Debug + Clone + Eq + Hash

"Interned" representation of a "generic parameter", which can be either a type or a lifetime. In normal user code, Self::InternedParameter is not referenced. Instead, we refer to Parameter<Self>, which wraps this type.

An InternedType is created by intern_parameter and can be converted back to its underlying data via parameter_data.

type InternedGoal: Debug + Clone + Eq + Hash

"Interned" representation of a "goal". In normal user code, Self::InternedGoal is not referenced. Instead, we refer to Goal<Self>, which wraps this type.

An InternedGoal is created by intern_goal and can be converted back to its underlying data via goal_data.

type InternedGoals: Debug + Clone + Eq + Hash

"Interned" representation of a list of goals. In normal user code, Self::InternedGoals is not referenced. Instead, we refer to Goals<Self>, which wraps this type.

An InternedGoals is created by intern_goals and can be converted back to its underlying data via goals_data.

type InternedSubstitution: Debug + Clone + Eq + Hash

"Interned" representation of a "substitution". In normal user code, Self::InternedSubstitution is not referenced. Instead, we refer to Substitution<Self>, which wraps this type.

An InternedSubstitution is created by intern_substitution and can be converted back to its underlying data via substitution_data.

type InternedProgramClauses: Debug + Clone + Eq + Hash

"Interned" representation of a list of program clauses. In normal user code, Self::InternedProgramClauses is not referenced. Instead, we refer to ProgramClauses<Self>, which wraps this type.

An InternedProgramClauses is created by intern_program_clauses and can be converted back to its underlying data via program_clauses_data.

type InternedProgramClause: Debug + Clone + Eq + Hash

"Interned" representation of a "program clause". In normal user code, Self::InternedProgramClause is not referenced. Instead, we refer to ProgramClause<Self>, which wraps this type.

An InternedProgramClause is created by intern_program_clause and can be converted back to its underlying data via program_clause_data.

type InternedQuantifiedWhereClauses: Debug + Clone + Eq + Hash

"Interned" representation of a list of quantified where clauses. In normal user code, Self::InternedQuantifiedWhereClauses is not referenced. Instead, we refer to QuantifiedWhereClauses<Self>, which wraps this type.

An InternedQuantifiedWhereClauses is created by intern_quantified_where_clauses and can be converted back to its underlying data via quantified_where_clauses_data.

type InternedParameterKinds: Debug + Clone + Eq + Hash

"Interned" representation of a list of parameter kind.
In normal user code, Self::InternedParameterKinds is not referenced. Instead, we refer to ParameterKinds<Self>, which wraps this type.

An InternedParameterKinds is created by intern_parameter_kinds and can be converted back to its underlying data via parameter_kinds_data.

type InternedCanonicalVarKinds: Debug + Clone + Eq + Hash

"Interned" representation of a list of parameter kind with universe index.
In normal user code, Self::InternedCanonicalVarKinds is not referenced. Instead, we refer to CanonicalVarKinds<Self>, which wraps this type.

An InternedCanonicalVarKinds is created by intern_canonical_var_kinds and can be converted back to its underlying data via canonical_var_kinds_data.

type DefId: Debug + Copy + Eq + Ord + Hash

The core "id" type used for struct-ids and the like.

type Identifier: Debug + Clone + Eq + Hash

Loading content...
Required methods

fn intern_ty(&self, ty: TyData<Self>) -> Self::InternedType

Create an "interned" type from ty. This is not normally invoked directly; instead, you invoke TyData::intern (which will ultimately call this method).

fn ty_data<'a>(&self, ty: &'a Self::InternedType) -> &'a TyData<Self>

Lookup the TyData from an interned type.

fn intern_lifetime(
&self,
lifetime: LifetimeData<Self>
) -> Self::InternedLifetime

Create an "interned" lifetime from lifetime. This is not normally invoked directly; instead, you invoke LifetimeData::intern (which will ultimately call this method).

fn lifetime_data<'a>(
&self,
lifetime: &'a Self::InternedLifetime
) -> &'a LifetimeData<Self>

Lookup the LifetimeData that was interned to create a InternedLifetime.

fn intern_parameter(&self, data: ParameterData<Self>) -> Self::InternedParameter

Create an "interned" parameter from data. This is not normally invoked directly; instead, you invoke ParameterData::intern (which will ultimately call this method).

fn parameter_data<'a>(
&self,
lifetime: &'a Self::InternedParameter
) -> &'a ParameterData<Self>

Lookup the LifetimeData that was interned to create a InternedLifetime.

fn intern_goal(&self, data: GoalData<Self>) -> Self::InternedGoal

Create an "interned" goal from data. This is not normally invoked directly; instead, you invoke GoalData::intern (which will ultimately call this method).

fn goal_data<'a>(&self, goal: &'a Self::InternedGoal) -> &'a GoalData<Self>

Lookup the GoalData that was interned to create a InternedGoal.

fn intern_goals<E>(
&self,
data: impl IntoIterator<Item = Result<Goal<Self>, E>>
) -> Result<Self::InternedGoals, E>

Create an "interned" goals from data. This is not normally invoked directly; instead, you invoke GoalsData::intern (which will ultimately call this method).

fn goals_data<'a>(&self, goals: &'a Self::InternedGoals) -> &'a [Goal<Self>]

Lookup the GoalsData that was interned to create a InternedGoals.

fn intern_substitution<E>(
&self,
data: impl IntoIterator<Item = Result<Parameter<Self>, E>>
) -> Result<Self::InternedSubstitution, E>

Create an "interned" substitution from data. This is not normally invoked directly; instead, you invoke SubstitutionData::intern (which will ultimately call this method).

fn substitution_data<'a>(
&self,
substitution: &'a Self::InternedSubstitution
) -> &'a [Parameter<Self>]

Lookup the SubstitutionData that was interned to create a InternedSubstitution.

fn intern_program_clause(
&self,
data: ProgramClauseData<Self>
) -> Self::InternedProgramClause

Create an "interned" program clause from data. This is not normally invoked directly; instead, you invoke ProgramClauseData::intern (which will ultimately call this method).

fn program_clause_data<'a>(
&self,
clause: &'a Self::InternedProgramClause
) -> &'a ProgramClauseData<Self>

Lookup the ProgramClauseData that was interned to create a ProgramClause.

fn intern_program_clauses<E>(
&self,
data: impl IntoIterator<Item = Result<ProgramClause<Self>, E>>
) -> Result<Self::InternedProgramClauses, E>

Create an "interned" program clauses from data. This is not normally invoked directly; instead, you invoke ProgramClauses::from (which will ultimately call this method).

fn program_clauses_data<'a>(
&self,
clauses: &'a Self::InternedProgramClauses
) -> &'a [ProgramClause<Self>]

Lookup the ProgramClauseData that was interned to create a ProgramClause.

fn intern_quantified_where_clauses<E>(
&self,
data: impl IntoIterator<Item = Result<QuantifiedWhereClause<Self>, E>>
) -> Result<Self::InternedQuantifiedWhereClauses, E>

Create an "interned" quantified where clauses from data. This is not normally invoked directly; instead, you invoke QuantifiedWhereClauses::from (which will ultimately call this method).

fn quantified_where_clauses_data<'a>(
&self,
clauses: &'a Self::InternedQuantifiedWhereClauses
) -> &'a [QuantifiedWhereClause<Self>]

Lookup the slice of QuantifiedWhereClause that was interned to create a QuantifiedWhereClauses.

fn intern_parameter_kinds<E>(
&self,
data: impl IntoIterator<Item = Result<ParameterKind<()>, E>>
) -> Result<Self::InternedParameterKinds, E>

Create an "interned" parameter kinds from data. This is not normally invoked directly; instead, you invoke ParameterKinds::from (which will ultimately call this method).

fn parameter_kinds_data<'a>(
&self,
parameter_kinds: &'a Self::InternedParameterKinds
) -> &'a [ParameterKind<()>]

Lookup the slice of ParameterKind that was interned to create a ParameterKinds.

fn intern_canonical_var_kinds<E>(
&self,
data: impl IntoIterator<Item = Result<ParameterKind<UniverseIndex>, E>>
) -> Result<Self::InternedCanonicalVarKinds, E>

Create an "interned" parameter kinds with universe index from data. This is not normally invoked directly; instead, you invoke CanonicalVarKinds::from (which will ultimately call this method).

fn canonical_var_kinds_data<'a>(
&self,
canonical_var_kinds: &'a Self::InternedCanonicalVarKinds
) -> &'a [ParameterKind<UniverseIndex>]

Lookup the slice of ParameterKind that was interned to create a ParameterKinds.

Loading content...
Provided methods

fn debug_struct_id(
struct_id: StructId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a type-kind-id. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_trait_id(
trait_id: TraitId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a type-kind-id. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_assoc_type_id(
type_id: AssocTypeId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a type-kind-id. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

fn debug_opaque_ty_id(
opaque_ty_id: OpaqueTyId<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of an opaque type. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific type-family (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_alias(alias: &AliasTy<Self>, fmt: &mut Formatter) -> Option<Result>

Prints the debug representation of an alias. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_projection_ty(
projection_ty: &ProjectionTy<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a ProjectionTy. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_opaque_ty(
opaque_ty: &OpaqueTy<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of an OpaqueTy. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_ty(ty: &Ty<Self>, fmt: &mut Formatter) -> Option<Result>

Prints the debug representation of an type. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_lifetime(
lifetime: &Lifetime<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of an lifetime. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_parameter(
parameter: &Parameter<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of an parameter. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_parameter_kinds(
parameter_kinds: &ParameterKinds<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a parameter kinds list. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_parameter_kinds_with_angles(
parameter_kinds: &ParameterKinds<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a parameter kinds list, with angle brackets. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_canonical_var_kinds(
canonical_var_kinds: &CanonicalVarKinds<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of an parameter kinds list with universe index. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_goal(goal: &Goal<Self>, fmt: &mut Formatter) -> Option<Result>

Prints the debug representation of an goal. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_goals(goals: &Goals<Self>, fmt: &mut Formatter) -> Option<Result>

Prints the debug representation of a list of goals. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_program_clause_implication(
pci: &ProgramClauseImplication<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a ProgramClauseImplication. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_program_clause(
clause: &ProgramClause<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a ProgramClause. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_program_clauses(
clauses: &ProgramClauses<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a ProgramClauses. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_application_ty(
application_ty: &ApplicationTy<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of an ApplicationTy. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_substitution(
substitution: &Substitution<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a Substitution. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_separator_trait_ref(
separator_trait_ref: &SeparatorTraitRef<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a SeparatorTraitRef. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

fn debug_quantified_where_clauses(
clauses: &QuantifiedWhereClauses<Self>,
fmt: &mut Formatter
) -> Option<Result>

Prints the debug representation of a QuantifiedWhereClauses. To get good results, this requires inspecting TLS, and is difficult to code without reference to a specific interner (and hence fully known types).

Returns None to fallback to the default debug output (e.g., if no info about current program is available from TLS).

Loading content...