Struct chalk_ir::Substitution

pub struct Substitution<TF: TypeFamily> {
    parameters: TF::InternedSubstitution,
}

A mapping of inference variables to instantiations thereof.

Fields
parameters: TF::InternedSubstitution

Map free variable with given index to the value with the same index. Naturally, the kind of the variable must agree with the kind of the value.

Methods

impl<TF: TypeFamily> Substitution<TF>

pub fn with_angle(&self) -> Angle<Parameter<TF>>

Displays the substitution in the form < P0, .. Pn >, or (if the substitution is empty) as an empty string.

impl<TF: TypeFamily> Substitution<TF>

pub fn from(
parameters: impl IntoIterator<Item = impl CastTo<Parameter<TF>>>
) -> Self

pub fn from_fallible<E>(
parameters: impl IntoIterator<Item = Result<impl CastTo<Parameter<TF>>, E>>
) -> Result<Self, E>

pub fn at(&self, index: usize) -> &Parameter<TF>

Index into the list of parameters

pub fn from1(parameter: impl CastTo<Parameter<TF>>) -> Self

pub fn empty() -> Self

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> Iter<Parameter<TF>>

pub fn parameters(&self) -> &[Parameter<TF>]

pub fn len(&self) -> usize

pub fn is_identity_subst(&self) -> bool

A substitution is an identity substitution if it looks like this

?0 := ?0
?1 := ?1
?2 := ?2
...

Basically, each value is mapped to a type or lifetime with its same index.