Module chalk_ir::visit

Traits for visiting bits of IR.

Re-exports

pub use visitors::VisitExt;

Modules

binder_impls
This module contains impls of Visit for those types that introduce binders.

boring_impls
This module contains "rote and uninteresting" impls of Visit for various types. In general, we prefer to derive Visit, but sometimes that doesn't work for whatever reason.

visitors
Visitor helpers

Enums

ControlFlow
An copy of the unstable std::ops::ControlFlow for use in Chalk visitors.

Traits

SuperVisit
For types where "visit" invokes a callback on the visitor, the SuperVisit trait captures the recursive behavior that visits all the contents of the type.

Visit
Applies the given visitor to a value, producing a visited result of type Visitor::Result.

Visitor
A "visitor" recursively folds some term -- that is, some bit of IR, such as a Goal, and computes a value as a result.