Trait chalk_ir::visit::Visit

pub trait Visit<I: Interner>: Debug {
    fn visit_with<'i, B>(
        &self, 
        visitor: &mut dyn Visitor<'i, I, BreakTy = B>, 
        outer_binder: DebruijnIndex
    ) -> ControlFlow<B>
    where
        I: 'i;
}

Applies the given visitor to a value, producing a visited result of type Visitor::Result.

Required methods

fn visit_with<'i, B>(
&self,
visitor: &mut dyn Visitor<'i, I, BreakTy = B>,
outer_binder: DebruijnIndex
) -> ControlFlow<B> where
I: 'i,

Apply the given visitor visitor to self; binders is the number of binders that are in scope when beginning the visitor. Typically binders starts as 0, but is adjusted when we encounter Binders<T> in the IR or other similar constructs.

Loading content...