Trait chalk_ir::zip::Zip

pub trait Zip<TF>: Debug where
    TF: TypeFamily, {
    fn zip_with<Z: Zipper<TF>>(
        zipper: &mut Z, 
        a: &Self, 
        b: &Self
    ) -> Fallible<()>;
}

The Zip trait walks two values, invoking the Zipper methods where appropriate, but otherwise requiring strict equality.

See Zipper trait for more details.

To implement the trait, typically you would use one of the macros like eq_zip!, struct_zip!, or enum_zip!.

Required methods

fn zip_with<Z: Zipper<TF>>(zipper: &mut Z, a: &Self, b: &Self) -> Fallible<()>

Loading content...