Trait chalk_ir::zip::Zipper

pub trait Zipper<TF: TypeFamily> {
    fn zip_tys(&mut self, a: &Ty<TF>, b: &Ty<TF>) -> Fallible<()>;

    fn zip_lifetimes(
        &mut self, 
        a: &Lifetime<TF>, 
        b: &Lifetime<TF>
    ) -> Fallible<()>;

    fn zip_binders<T>(&mut self, a: &Binders<T>, b: &Binders<T>) -> Fallible<()>
    where
        T: Zip<TF> + Fold<TF, TF, Result = T>;
}

When we zip types, we basically traverse the structure, ensuring that it matches. When we come to types/lifetimes, we invoke the callback methods in the zipper to match them up. Primarily used during unification or similar operations.

So e.g. if you had A: Eq<B> zipped with X: Eq<Y>, then the zipper would get two callbacks, one pairing A and X, and the other pairing B and Y.

For things other than types/lifetimes, the zip impls will guarantee equality. So e.g. if you have A: Eq<B> zipped with X: Ord<Y>, you would wind up with an error, no matter what zipper you are using. This is because the traits Eq and Ord are represented by two distinct ItemId values, and the impl for ItemId requires that all ItemId in the two zipped values match up.

Required methods

fn zip_tys(&mut self, a: &Ty<TF>, b: &Ty<TF>) -> Fallible<()>

Indicates that the two types a and b were found in matching spots, beneath binders levels of binders.

fn zip_lifetimes(&mut self, a: &Lifetime<TF>, b: &Lifetime<TF>) -> Fallible<()>

Indicates that the two lifetimes a and b were found in matching spots, beneath binders levels of binders.

fn zip_binders<T>(&mut self, a: &Binders<T>, b: &Binders<T>) -> Fallible<()> where
T: Zip<TF> + Fold<TF, TF, Result = T>,

Zips two values appearing beneath binders.

Loading content...